+++ tests/end_to_end/tests.py
@@ -25,8 +25,9 @@ from xrdsst.core.conf_keys import ConfKeysSecurityServer
 from xrdsst.core.definitions import ROOT_DIR
 from xrdsst.core.util import revoke_api_key, get_admin_credentials, get_ssh_key, get_ssh_user
 from xrdsst.main import XRDSSTTest
+from xrdsst.models import ClientStatus, CertificateDetails
+from xrdsst.models.key_usage_type import KeyUsageType
 from xrdsst.models import ClientStatus
-
 
 class EndToEndTest(unittest.TestCase):
     config_file = None
@@ -936,10 +937,64 @@ class EndToEndTest(unittest.TestCase):
             assert len(certificates) == 6
             assert len(cert_controller.app._last_rendered[0]) == 7
 
+    def step_disable_certificates(self):
+        with XRDSSTTest() as app:
+            cert_controller = CertController()
+            cert_controller.app = app
+            cert_controller.load_config = (lambda: self.config)
+
+            certificates = cert_controller.list()
+
+            for security_server in self.config["security_server"]:
+                security_server["certificate_management"] = [cert["hash"] for cert in certificates if cert["ss"] == security_server["name"]]
+
+            cert_controller.load_config = (lambda: self.config)
+            cert_controller.disable()
+            certificates_disabled = cert_controller.list()
+            for cert_disabled in certificates_disabled:
+                assert cert_disabled["ocsp_status"] == "DISABLED"
+
+    def step_unregister_certificates(self):
+        with XRDSSTTest() as app:
+            cert_controller = CertController()
+            cert_controller.app = app
+            cert_controller.load_config = (lambda: self.config)
+
+            certificates = cert_controller.list()
+
+            for security_server in self.config["security_server"]:
+                security_server["certificate_management"] = [cert["hash"] for cert in certificates
+                                                                  if cert["ss"] == security_server["name"] and cert["type"] == KeyUsageType.AUTHENTICATION]
+
+            cert_controller.load_config = (lambda: self.config)
+            cert_controller.unregister()
+            certificates_unregister = cert_controller.list()
+            for cert_unregister in certificates_unregister:
+                if cert_unregister["type"] == KeyUsageType.AUTHENTICATION:
+                    assert cert_unregister["status"] == "DELETION_IN_PROGRESS"
+
+    def step_delete_certificates(self):
+        with XRDSSTTest() as app:
+            cert_controller = CertController()
+            cert_controller.app = app
+            cert_controller.load_config = (lambda: self.config)
+
+            certificates = cert_controller.list()
+
+            for security_server in self.config["security_server"]:
+                security_server["certificate_management"] = [cert["hash"] for cert in certificates
+                                                                  if cert["ss"] == security_server["name"] and cert["type"] == KeyUsageType.AUTHENTICATION]
+
+            cert_controller.load_config = (lambda: self.config)
+            cert_controller.delete()
+            certificates = cert_controller.list()
+            for cert in certificates:
+                assert cert["type"] != KeyUsageType.AUTHENTICATION
+
     def test_run_configuration(self):
         unconfigured_servers_at_start = self.query_status()
 
-        self.step_member_find()
+        # self.step_member_find()
         self.step_verify_initial_transient_api_keys()
         self.step_upload_anchor_fail_file_missing()
         self.step_upload_anchor_fail_file_bogus_content()
@@ -1008,3 +1063,7 @@ class EndToEndTest(unittest.TestCase):
         configured_servers_at_end = self.query_status()
 
         assert_server_statuses_transitioned(unconfigured_servers_at_start, configured_servers_at_end)
+
+        self.step_disable_certificates()
+        self.step_unregister_certificates()
+        self.step_delete_certificates()